Added test for sending and receiving data using sockets
Moved the server creation to a separate function. Since the server runs on the main process, we could only create new threads or processes when the server received data. Moving the server into a separate function allows for us to unit test the server, and to unit test sending data to the server.

Added send_data_socket function to the xcash_proof_of_stake_functions_and_macros.c file. This function will create a client, send data through a TCP port to a server, and then close the socket. This is what we will run on separate threads to send data, while our server runs on a separate process per connection.

Added create_server_on_separate_thread to the xcash_proof_of_stake_test.c file. This is just a thread wrapper function that will call the server function on a different thread for the unit testing.

Added the unit test for creating the server and sending and receiving data using sockets, since this is the main part of the system.

Added descriptions to some of the functions that did not have descriptions.