Upgraded string_replace and created sign and verify data functions
Upgraded string_replace since it was not working when a string had a substring in it. It was also not replacing all occurrences of the string to replace.

Added unit testing for the string_replace function.

Added random_string, get_public_address, sign_data, verify_data functions. These functions will be used by the client to securely send a message and the server to securely receive a message, so there is no possibilities for man in the middle attacks. The random_string function will change the signature of the message each time, even if the message is the same, to prevent frequency analysis or spoofing attacks.

Added unit testing for get_public_address, sign_data, verfiy_data functions.

Created a unit test for sending signed data to server, then the server verifies the data and then sends signed data to the client, who then verifies the data.

Added server_received_data_xcash_proof_of_stake_test_data since each type of received message from the server will be processed in a separate function.