Increasing security for messages
Cleaned up some variable names, and added comments when variable parameters were missing.

Made every function const, all capitals.

Added a pointer_reset_all macro to each function to make the function easier to read.

Added get_current_block_height function to get the current block height.

Changed some constants to define macros in the random_string function.

Added the current block height, current part of round and current backup node for the current part of the round to the sign_data and verify functions. These 3 items in combination with the random character string and the signing of the data with the nodes public XCA address, will stop all spoofing, man in the middle, or replay attacks on messages. This also allows the nodes to not even need to know the IP from the consensus node, since the IP will not be needed in the every 24 hour recalculation.