Added string_count function
Added string_count function to count occurrences of a substring in a string.

Added string_count_test function to test the string_count function.

Fixed a memory leak in string_replace.

Moved the Global Structures and Global Variables above the Function prototypes since the other way would not work if trying to use a global struct in a function.