Created SERVER_ERROR message
Created SERVER_ERROR message define macro to close the client socket, reset all of the variables on the heap, and either return 0 from the function or exit the forked process. The reason for creating the define macro was so the code is easier to read in the server message part of the create_server function.