Changed the NODES_TO_CONSENSUS_NODE_VOTE_RESULTS message format
Changed the NODES_TO_CONSENSUS_NODE_VOTE_RESULTS message format since this will need to return the main node data to the consensus node, so it can give this data to the block producer.

Added current_round_part_consensus_node_data struct to hold all of the main node data to send to the consensus node for each part of the round.