Upgraded VRF creation method
Upgraded the VRF creation method to be more secure by using all block verifiers to create the alpha string, and to create the VRF secret key and VRF public key before the alpha string is created, this way block verifiers can not create a favorable VRF beta string.

In doing this, it also reduces the bandwidth needed for each round, and reduces the minimum time a round can take, which will be beneficial for side channels and side chains.