Added append_string define macro function
Added append_string define macro function, that will automically copy a string at the end of the other string. It will also get the length of the first string and the total size of the first string and will not allow more bytes than the total size of the first string.