Added block verifiers time sync check
Added a check to make sure that a block verifier is within 2 seconds of the current time of the network data nodes. This will ensure that with the network latency and differences in time, all block verifiers will be within the 2 second allowed difference before they are able to be a block verifier.