Added get_delegates_online_status function and increased stack size
Added get_delegates_online_status function to get all of the delegates online status before checking if the database is synced.

Added a check to not start any of the timers until enough block verifiers are online in the database.

Increased the stack size and file limits to infinity for the systemd service file.