Fixed security issue reserve proofs
Fixed a security issue with reserve proofs, where if someone has access to the network data nodes, they could change the total votes for users, and it would not have been detected since the reserve proof is still valid, and the amount is calculated when the reserve proof is registered into the database. Fixed the issue by not only checking if the reserve proof is invalid, but also checking if the amount in the database is the same amount that is returned from the check_reserve_proofs function, that gets the amount from the blockchain.