Changed gmtime to gmtime_r
Changed gmtime to gmtime_r, since it could have been causing problems with accepting messages in the round. This is because gmtime is not re-entrant and since a timer is constantly checking if the current minute is a divisible by 5 minute this could have been erasing the result of gmtime in the round to check if certain messages are received between certain times.

Replaced in the server receive messages part of the code the inet_ntop, inet_aton and gethostbyaddr systemcalls to getnameinfo since gethostbyaddr is not re-entrant.