Added sync from main network data node if no consensus
Added a check after syncing, to check if your DB is in the majoirty, and if not sync directly from the main network data node. This if no majority can be reached, or the block verifier synced from an out of date node, they will sync from the main network data node and a majority will be able to be reached.