Removed verify message when users sync the database
Removed verify message when users sync the database, this is due to the fact that they will not have there ECDSA key in the DB yet since they are not registered, so the block verifier will have no way of being able to verify the message. Did not change the verify message for these messages back to the public XCASH address, since the DB is decentralized anyone should be able to download the DB, so the verify message function was removed for the download, but kept for the when the user verifies from the block verifier they have a correct DB from a verified source.