Fixed check what block verifiers key to use when validating a block.
Added the check to make sure that the selected block verifier (whos public and secret key will be used for the VRF data) did not timeout or not provide a valid key. This check was in the code when creating the VRF data, but not in the code when verifiying a block is correct.