Added live syncing mode
Added support for live syncing mode. The function will currently just figure out how many blocks reserve bytes database hash it needs to send to the node to be fully synced. It however will return 0 when live syncing (a node that is fully synced and syncs the blocks one by one as they come in). So if it returns 0 we just set it to 1 to enable live syncing mode and still keep the valid amount of blocks for a node that needs to sync a lot of blocks.