Changed error on block submit function
Changed what happens if there was an error on block submit. The block is valid if it reaches that point, meaning the block will be added to the network. Instead of restarting the round if your block verifier does not see it added to the network, it should let the block producer submit the block, and then the block producer does not, then have each network data node submit the block to make sure it gets added, and then have all block verifiers go to the next round.