Changed memory allocaion for invalid reserve proofs
Changed memory allocaion for invalid reserve proofs, instead of allocating the maximum amount at the start of the program, it now will allocate one reserve proof right before it adds the reserve proof to the invalid_reserve_proofs struct. This might make this process an unnoticeably amount slower, but it saves an estimated 1.5 GB of memory depending on how many invalid reserve proofs are found during the round.