Fixed syncing db at start of the round
Fixed syncing db at start of the round, as it was only syncing the reserve bytes not all db if not in sync, and it was not returning a status that synced the db, so it would not update the previous block hash after syncing. This then caused an infinite loop where a block verifier could not get back into participating in the rounds.