Stable mining and reserve proofs check test
At this point from a day test the code seems stable with mining while reserve proof checking is ongoing.

Changes made to the code were to move the reserve proof checking after a round is complete and while it is waiting for the 5 minute interval. This was because since the xcash-core is not multithreaded, constant RPC calls to these would make the mining process not work since they rely on them to get a block template submit a block template etc, or have a higher rate of failure for some delegates.

Becuase of this voting is now only allowed on the first block of every hour since the reserve proof checking has a limited amount of time. Its estimated that now about 60000+ reserve proofs are checked within the hour. Votes are still able to be cancelled at any time and will update at the next block, but a revote will have to wait for the next hour to avoid double voting.

A follow up test will be conducted for a longer period of time after LXC container test and a test of the websites while mining.