fix: Fixed keeping the network data nodes synced
Fixed keeping the network data nodes synced, because if this sync check ran while waiting for the block producer to submit the block, then if a network data node was not synced, it would never run that function, so it needed to be moved to the start of the round.

Speed up database syncing for block verifiers in between rounds.

Removed the second backup block producer to make time for the syncing of both the network data nodes and the block verifiers at the begining of the round. This means that there is only the block producer and a backup block producer before the network data node will create the block.