fix: dont retry db syncing if from specific delegate
If syncing the database from a specific delegate, if an error occurs just exit, dont keep looping and trying to connect to the same delegate. This is why this function is only used as a quick bootstrap option at the begining of the program, and is only used when there is no consensus among network data nodes, since at that point they need to sync from the main network data before continuing any block mining.