feat: Optimized VRF data from O2 to O
Optimized VRF data from O2 to O, by making sure that the block verifiers add the other block verifiers signatures in order of the current block verifiers list. This still needs to be checked when delegates change ranks between rounds but should be fine.

In doing so, this has optimized how many calls needed to VRF_data_verify from 10000 to 100, and this will mostly be noticed when tokens are mainnet, as this will reduce each sidechains calls to this function by squared factor. This also allowed for some more time in the round to be added to the block producer creating and sending the block to the other block verifiers, and thus will drop the bandwidth needed by the block producer to around 8 MBPS.

Changed the ECDSA block verifiers key, so each network data node will have their own key for production, and each block verifier in test mode will use a different key.

Removed uneeded test due to the O2 to O optimization