fix: vote total not increasing after vote was made
Fixed the reason why the delegates vote count was not increasing when a vote was being made.

Updated the unit test to check for if a delegates vote count is properly increased when a voter votes for it, and checked for when a voter votes for another delegate, does the previous delegates vote count decrease and does the new delegates vote count increase, even if its a revote for the exact same delegate just with a different amount, and does it not affect any other voters amoutns who voted for the delegate.