Fix: check if network data nodes have the previous blocks reserve bytes when syncing
If there is not a majority among the seed nodes, it needs to check if the main seed node has the previous blocks reserve bytes in the database and if not sync from the second one or third one etc etc. This is due to a bug that occured in the alpha test, where the main seed node missed the current block, but others node mined it, then the next round the network data nodes could not form a majority and synced from the main network data node, thus removing the previous blocks reserve bytes from there database and everyone elese database, thus the chain could not continue.