fix: recalculate the delegates if your db is unsynced
Added a function to recalculate only the next block verifiers, if your db was unsycned and you synced it from a block verifier. This is due to the fact that if invalid reserve proofs were found in the round it will be working on a incorrect total vote count and have an incorrect delegates order.