fix: Change sync settings if no majority
If there is no majority, change the sync settings to on for the network data node the other network data nodes are going to sync from, and then after they are done syncing, set there sync settings to on. Two errors were caused as none of the network data nodes could sync from the main one when there was not a majority, and after a round that did not have a majority any block verifier could not sync from any of the network data nodes.