fix: Changed how the first or only network data node is detected
Changed how the first or only network data node is detected for when syncing the previous, current and next block verifiers and when seeing if the databases need to be checked if they are synced at the start of the program. Originally it was just checking if you were the main network data node, however that does not work for either case and it was changed to if your the only network data node online.