fix: detection if delegate was online
Fixed function that checked if a delegate was online, it was only checking if the server was online and the port was open, but not if the XCASH DPOPS program was actively running. Also changed the block verifiers total rounds so they now will not show 100 percent reliablity for delegates that were in the top 100 and were offline for that round.