fix: Added limits where 50MB allocation
Added limit to number of open connections where 50MB allocation for server functions. The limit only allows an IP address one open connection at a time, and will on top of verifying the message was signed by the correct public address, it will also require that the public address is a registered delegate and that the delegate has over 100 million xcash. This is due to the fact that the syncing is when a delegate is already in the top 100, or lower than the top 100 and leaves the program running so it is syned if they become the top 100. This should block most attacks, with the barrier of multiple IP addresses, and having a registered delegate with 100 million xcash, and having something to loose if a delegate by others not voting for it. This is another reference to issue #17