feat: Added anti DDOS system
Added a anti DDOS system. The premise of the anti DDOS system is to prevent regular flood style DDOS attacks like referenced in issue #16 and resource style attacks like mentioned in issue #17. These are very different attacks, as one uses a high amount of connections to flood a server, and the other one uses long timeouts etc to block others from being able to connect since it will try to use all of the available connections.

To combat these two attack styles we will use a system of limiting the number of simultaneous connections an IP address can have open, and the other being limiting the number of simultaneous connections one block verifier can have open to another block verifier. The reason that this works better than just limiting IP addresses, is that IP addresses can be spoofed, but block verifier messages can not be spoofed, since they are signed with the block verifiers ECDSA key. The goal is to try to use this type of connection limit wherever possible, and if not possible use the IP address limits. We also want to limit people spamming the DB to become a registered delegate to DDOS, so ontop of this the check for a valid registered delegate will also need to have 100 million xcash to send a message. This works because for most messages in the system, only the top 100 delegates should be communicating with each other, and maybe the next 100 who are live syncing but not participating.

The one problem with this system of limiting either the public address of a block verifier or the IP address of a message to a maximum amount of connections is the xcashd needs to send a message to other block verifiers to sync. Since all block verifiers IP addresses are public, a situation can arise where an attacker could spoof messages from a block verifiers IP address, and not allow their xcashd to send messages to other block verifiers. To combat this, we have added parameters to the xcashd to add in your block verifiers public address and ECDSA secret key, this way the xcashd can now sign messages just like the XCASH_DPOPS program. If the parameters are not run, the xcashd messages will be limited by IP address since these will be normal users who are not block verifiers, and if the parameters are used, these will be limited on the public address, since the IP is known and could be spoofed but the ECDSA secret key can not be.

To summarize these are the following DDOS rules for these types of connections:

Any message between a block verifier and a block verifier would require a signed message with the public key and 100M xcash and limited to 20 connections per public address
Any message for the http server is limited to 20 connections per IP address
Any message from xcash-wallet-cli or xcash-wallet-rpc is limited to 20 connections per IP
Any message from xcashd is either required to have a signed message with the public key and 100M xcash and limited to 20 connections per public address, or is limited to 20 connections per IP.