fix: try to solve random not receive message bug
Various changes to try to fix the bug of some nodes not receiving data randomly.

Changed all memcmp to strncmp to avoid an out of bounds memory check if the hard coded string is longer than the other.

Removed the double call of getnameinfo in the server loop, as this was added during the DDOS and might be a problem.