autoinstaller.sh: remove previously downloaded bootstrap-file added
Issue:
Interrupting the download of a bootstrap-file leaves an incomplete file on the server.
When the download process is repeated wget will save the new download as "original name + index".
When unpacking the bootstrap-file 7zip will try to unpack the incomplete file downloaded the first time and exit with an error.   

The added code removes eventually existing previously downloaded copies of the bootstrap.