fix: Added better check if blockchain is not synced
Added better check if blockchain is not synced, checking for block height, and for if xcash daemon is busy with no in peers yet. Sometimes this can happen on a new fork, you have not ran yet, and it can sometimes take 10 minutes to find an in peer. If your not fully connected the XCASH_DPOPS would loop foreever until you restarted it.