fix: Various stability fixes
Fixed an issue with the server could sometimes not accept connections or could add mulitple file descriptors for the same connection, or could close the connection but still have it in the file descriptors.

Added restart if your server gets stuck.

Added either 67% or 51% with 4/5 seed nodes. Might not use this, but its good to have if we need to use it the first week or something.