fix: rewrote update_block_verifiers_list
rewrote update_block_verifiers_list, as delegates that got off sync would sync the database the next round, but would still be off on the order of the delegates. Rewrote it instead of just getting the current list by vote count for the next block verifiers and moving the next to current and current to previous, The current and previous are now loaded directly from the reserve bytes in the previous and previous block before the previous block, so they will all have the same order.