fix: Removed update block verifiers before syncing database
Removed update block verifiers before syncing database, as with changing how the block verifiers are loaded, this would cause an out of sync delegate to get stuck every round.

Changed the update statistics to use the current block verifiers since we are not updating until after the database is synced, so these are really the previous block verifiers.