fix: better checking if delegate was previous block prouducer
Added better checking for if the delegate was the previous block producer, instead of using a variable to hold the previous block producer, the previous block producer is now read from the database from the last block. This ensures that all delegates have the same results. This also fixes where sometimes it would say a delegate was the previous block producer, when it was not. This also fixes where sometimes it would not check if it was the previous block producer even though it was.

Changed the time of when the timer runs from right after the block is created, to the end of the database syncing time, this way all delegates have the same database to read the previous block producer from.