fix: Changed double for loop valid conditions blockchain signatures.
Changed double for loop valid conditions blockchain signatures, as before if it failed to check the signatures in order, it would check all of them in a double for loop. however it required not only 67% but 4/5 network data nodes to have signed the block. This is not a protocol rule for xcash-core so just setting it to 67% is enough.