fix: dont allow seed nodes to mine blocks
Dont allow the seed nodes to mine blocks. The whole point of them is to keep give weight to the vote count, and integrity of the decentralized db, and the on chain db. So because of this they dont really need to be the block producer, only unless all bytes of the VRF beta string run out, which should never to very rarely happen.