fix: Change replay round seed
Changed the replayed round seed from the VRF beta and DB hash to the VRF beta and the current time. This should fix two issues.

The first issue was that some users could have different db hashes if lots of people are updating data so that would mean different block producers.

The second issue sometimes the database will not change between 5 minutes so in a replayed round it would select the same two block producers.