fix: Buffer overflow with reserve proof vote count
A buffer overflow can occur since when adding a reserve proof to the database we are allowing RPC error response to be valid, thus retruning an empty string for the total vote and having a buffer overflow. We only want to add the reserve proof to the database if its valid.