Add the option (to the user) to chose between Opacity Handle or direct URL for image download + change script URLS pointing to X-Cash GitHub
The script asks the user if he wants to use Opacity File Handle (and so download using Opacity Storage downloader) or use a direct download of the image from a URL (using wget). Opacity is an anonymous crypto cloud storage where files are encrypted/decryped client side. Each file handle is pointing to a unique file, so a new image version will have a new file handle. Currently Opacity Network is not decentralized (it's like traditional cloud but without asking personal info, with cheap storage payable with OPQ token) and can be really fast for downloading files. 
If the user choose to download from URL and the script contains a default download URL (eg. official mirror) it will be asked if use that one. There is always the possibility to enter a custom one (that must be a tar.gz of an X-Cash pre-built image).
Currently the default option (if the user press simply enter) is to use Opacity and ask for a download handle.

Since the script is now part of the official X-Cash repository, the URLs of the scripts (variables inside the lxc_autoinstaller) has been updated, pointing to `X-CASH-official/xcash-dpops` repository.