fix: error with false postive for replayed round
error with false postive for replayed round, and a non replayed rounds reserve bytes would be removed.

This was due to the decentralized syncing taking longer and the delegate checking the N-2 block instead of the N-1 block.

Instead of removing the block at the start of the round, remove the block if it exist in the database before you add the "double" block.