fix: Replayed round edge cases
Fixed various replayed round edge cases.

Added a false postive check for replayed round. This way if it is a false positve it will sit out the round, but if its a real replayed round and there is an invalid blocks data in the database, it will remove it before adding the current block.

A false postive for a replayed round can occur when the new round starts, but the blockchain height has not synced the current block height yet. In this case we check again 2 minutes later after the db has been synced.

A replayed round can occur when the block can not be added to the blockchain and the xcash DPOPS round is replayed.There is no extra steps since nothing was added to the blockchain or database.

A replayed round that has invalid database data can occur when there was enough block verifiers to validate the block and add it to the database in xcash DPOPS, but later there is not enough block verifiers to add the block to the blockchain.In this case we remove the invalid blocks data before we add the new blocks data.