fix: Updates to autoinstaller
Changed source commands to always end true, for non root users using the new way of running the script.

Load the systemd and firewall files dynamically, so they dont have to be included in the autoinstaller.

Added setup options for a Node only to the menu (have not added the code yet)