fix: lots of no majoirty could be reached
It seems a lot of times maybe 15% of the time it shows no majority could be reached. This should rarely happen, so increase the time by 10 seconds, and change the sleep to sync in syncing function as some might have checked the results before the rest of the nodes could send there results.