feat: 0 network data nodes
Add code to allow the xcash DPOPS to function with 0 network data nodes running.This will increase the DDOS chances substantially since 55% or 67% of nodes will have to be online for block producing to stop. The only thing that wont work on the network is any wallet functions if there are 0 network data nodes online, and only block verifiers will be able to sync the blockchain, no other user can sync the blockchain. This fix is meant to temporary keep the block producing going until a network data node can get back online.