fix: various optimizations
if 0 seed nodes are online, and the other block verifiers create blocks, there was currently no way for the seed nodes to turn back on, since the xcash core would be stuck.

The steps to turn on the seed nodes:

select one seed node.
restart the xcash dpops.

restart the xcashd and xcash wallet rpc.

restart the xcash dpops.

For the other seed nodes, just use the program restart in the autoinstaller.