fix: reserve proofs data hash always empty string
The reserve proofs were checked by checking all 50 reserve proofs databases. If any of the 50 were empty, it would give the entire data hash as the empty string. It now will only check the reserve proofs that are not empty and get the real data hash.

This had no affect on the code before since we checked all the databases at once, but now since were checking them individually it would have always showed the reserve proofs are synced with the majority.