feat: split db sync to individual db
Split the db syncing instead of all of the db if not in the majority, to checking if each individual db is in the majority and only syncing the ones that are needed. This will save extra bandwidth and will not get extra false positives on the new reserve proof checking.