fix: Check for individual and overall majority
Since XCASH DPOPS uses a custom built DBFT consensus mechanism, before we were only checking for an overall majority (VRF data, block template, block template signature, reserve bytes). However this could lead to a malfunctioning node messing up the majority count.

To fix this still check an overall majority for every data received, but before this check, check for an individual majority for each block verifier. This allows for a malfunctioning node to be set to the empty string result, instead of accepting whatever random data it could have sent to many different block verifiers. This also allows the program to print out any block verifier that is malfunctioning, so the other block verifiers know that it is malfunctioning.
This will double the amount of messages sent during the round and increase the bandwidth since the timings have been decreased to add these extra checks.

Adjusted the timings.

Added a new GUI printout for the rounds.