fix: stats db was not organized
The statistics database was not organized, due to the way it was finding the best block verifier for the statistics. This could have lead to some block verifiers not having the majority when syncing the database, and lead to the stats page on the website to be wrong.The statisitics are now selected from the organized list of block verifiers so it will choose the best block verifier on the list that is towards the top of the list.