feat: DB integrity check
Added a DB integrity check to run once every 24 hours. This is more for a on paper theory, but essentially 27+ delegates can control the network which is normal in DBFT. However they cant do much like mine blocks or something as that would destroy the coin. However they could shift votes or add to vote totals and have it not be noticed. This system resets that behaviour by making sure all reserve proofs add up to the delegates total_vote_count.

This is also a protection from rounding or buffer overflows, as well, since the system would not overflow, but leave a postive amount with no reserve proof backing it, if an overflow occured.