feat: Update reserve proofs db for 1M vote and fix error
Fixed an error in the reserve proofs db, where it should have been using a second collection per 1000 reserve proofs. However it was checking not all reserve proofs so it would not return the amount, and we have now gone over 1000.

Fixed it so we now can do 2000 as 1000 was only 4mb, but also using this to change the db so it can hold 1M vote now (we just need to change the vote limit from 2m to 1m).

Removed about 0.2mb of waste on the stack.