fix: Fix log print out for IP address was cut at sizeof(pointer)
when I changed delegates and statistics var from stack to heap, I forgot to change max size protection for them, so it was thinking sizeof(char*) was the max it could hold, and cutting everything off at 8 bytes since that is the size of a memory address on a 64bit system. changed this to BLOCK_VERIFIERS_IP_ADDRESS_TOTAL_LENGTH that is 100 bytes so it can display the full ip.

do note this bug does not affect any functionality since it was just cutting the writing function off at 8 bytes.