(dont download) Temporary fix test of individual majority
DONT DOWNLOAD STILL TESTING

Testing a temporary fix for the individual majority. It works correctly but for some reason not enough servers can send this data (about 12-20) and their not sending it late.

Once this happens, it switches everyone data to 0's, and then thinks that no one sent any data.

This is also a smaller problem we have seen since launch, maybe this will fix this as well.

As to why this happens here and their and is happening constantly now is still to be determined.