feat: Add delegate error reporting
Added delegate error reporting system.

This system can be turned on with adding the debug-delegates-error parameter.

This will report what delegates are not responding for each specific part of the round (except for part 25 still need to try to figure this part out).

This should allow the delegates as a group to message one another if they are showing up round after round so all delegates can be more stable.

This fixes #64 from the community request